Переработка функции отрисовки элементов. Разделение utils на несколько файлов
