+++ src/mock-data/movies-data.js
@@ -1,14 +1,8 @@
 import {getRandomCommentsData} from './comments-data';
 import {
-  getRandomBoolean,
   getRandomDate,
-  getRandomDescription,
-  getRandomFloat,
-  getRandomInteger,
-  getRandomItem,
-  getRandomSubArray,
   transformDuration
-} from '../services/utils';
+} from '../utils/date';
 import {
   Actors,
   Descriptions,
@@ -19,6 +13,14 @@ import {
   ReleaseCountries,
   Writers
 } from './data-sets';
+import {
+  getRandomBoolean,
+  getRandomDescription,
+  getRandomFloat,
+  getRandomInteger,
+  getRandomItem,
+  getRandomSubArray
+} from '../utils/common';
 
 const MIN_COMMENTS_NUMBER = 0;
 const MAX_COMMENTS_NUMBER = 5;
