+++ src/modules/render-films-number.js
@@ -1,8 +1,8 @@
 import NumberOfFilms from '../view/number-of-films';
-import {insertDOMElement, Positions} from '../services/utils';
+import {renderDOMElement, Positions} from '../utils/render';
 
 export const renderFilmsNumber = (data) => {
   const numberOfFilmsContainer = document.querySelector('.footer__statistics');
-  const numberOfFilms = new NumberOfFilms(data).getElement();
-  insertDOMElement(numberOfFilmsContainer, numberOfFilms, Positions.BEFOREEND);
+  const numberOfFilms = new NumberOfFilms(data);
+  renderDOMElement(numberOfFilmsContainer, numberOfFilms, Positions.BEFOREEND);
 };
