+++ src/modules/render-sort-panel.js
@@ -1,8 +1,8 @@
 import SortPanel from '../view/sort-panel';
-import {insertDOMElement, Positions} from '../services/utils';
+import {renderDOMElement, Positions} from '../utils/render';
 
 export const renderSortPanel = () => {
   const mainContainer = document.querySelector('.main');
-  const sortPanel = new SortPanel().getElement();
-  insertDOMElement(mainContainer, sortPanel, Positions.BEFOREEND);
+  const sortPanel = new SortPanel();
+  renderDOMElement(mainContainer, sortPanel, Positions.BEFOREEND);
 };
