+++ src/modules/render-user-rank.js
@@ -1,8 +1,8 @@
-import {insertDOMElement, Positions} from '../services/utils';
 import UserRankView from '../view/user-rank';
+import {renderDOMElement, Positions} from '../utils/render';
 
 export const renderUserRank = (data) => {
   const userRankContainer = document.querySelector('.header');
-  const userRank = new UserRankView(data).getElement();
-  insertDOMElement(userRankContainer, userRank, Positions.BEFOREEND);
+  const userRank = new UserRankView(data);
+  renderDOMElement(userRankContainer, userRank, Positions.BEFOREEND);
 };
