+++ src/modules/show-popup-handler.js
@@ -1,7 +1,8 @@
-import {getRandomInteger, insertDOMElement, isEscEvent, Positions} from '../services/utils';
 import {moviesData} from '../mock-data/movies-data';
 import CommentItem from '../view/popup-comment';
 import FilmPopup from '../view/film-popup';
+import {renderDOMElement, Positions} from '../utils/render';
+import {getRandomInteger, isEscEvent} from '../utils/common';
 
 export const setOpenPopupHandler = () => {
   const cardsContainers = document.querySelectorAll('.films-list__container');
@@ -16,14 +17,14 @@ export const setOpenPopupHandler = () => {
       //todo: Реализовать подстановку данных того фильма, на который кликнули, вместо случайного
       const movieItem = moviesData[getRandomInteger(0, moviesData.length - 1)];
       const {comments} = movieItem;
-      const filmPopup = new FilmPopup(movieItem).getElement();
+      const filmPopup = new FilmPopup(movieItem);
 
       if (
         clickEvt.target.classList.contains('film-card__poster') ||
         clickEvt.target.classList.contains('film-card__title') ||
         clickEvt.target.classList.contains('film-card__comments')
       ) {
-        insertDOMElement(popupContainer, filmPopup, Positions.AFTEREND);
+        renderDOMElement(popupContainer, filmPopup, Positions.AFTEREND);
         document.body.style.overflow = 'hidden';
         const popup = document.querySelector('.film-details');
         const closeButton = popup.querySelector('.film-details__close-btn');
@@ -48,7 +49,7 @@ export const setOpenPopupHandler = () => {
 
         comments.forEach((commentItem) => {
           const comment = new CommentItem(commentItem).getElement();
-          insertDOMElement(commentsContainer, comment, Positions.BEFOREEND);
+          renderDOMElement(commentsContainer, comment, Positions.BEFOREEND);
         });
       }
     });
