+++ src/view/abstract-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../services/utils';
+import {createElement} from '../utils/render';
 
 export default class AbstractView {
   constructor() {
